Moved getColorResource and getDrawableResource to ResourceUtils
