+++ WordPress/src/main/java/org/wordpress/android/ui/reader/ReaderSubsActivity.java
@@ -43,6 +43,7 @@ import org.wordpress.android.util.AppLog;
 import org.wordpress.android.util.DisplayUtils;
 import org.wordpress.android.util.EditTextUtils;
 import org.wordpress.android.util.NetworkUtils;
+import org.wordpress.android.util.ResourceUtils;
 import org.wordpress.android.util.ToastUtils;
 import org.wordpress.android.util.UrlUtils;
 import org.wordpress.android.widgets.WPViewPager;
@@ -89,8 +90,8 @@ public class ReaderSubsActivity extends AppCompatActivity
 
         TabLayout tabLayout = (TabLayout) findViewById(R.id.tab_layout);
         tabLayout.setTabMode(TabLayout.MODE_SCROLLABLE);
-        int normalColor = ReaderUtils.getColorResource(this, R.color.blue_light);
-        int selectedColor = ReaderUtils.getColorResource(this, R.color.white);
+        int normalColor = ResourceUtils.getColorResource(this, R.color.blue_light);
+        int selectedColor = ResourceUtils.getColorResource(this, R.color.white);
         tabLayout.setTabTextColors(normalColor, selectedColor);
         tabLayout.setTabMode(TabLayout.MODE_SCROLLABLE);
         tabLayout.setupWithViewPager(mViewPager);
