+++ WordPress/src/main/java/org/wordpress/android/ui/reader/utils/ReaderUtils.java
@@ -2,12 +2,8 @@ package org.wordpress.android.ui.reader.utils;
 
 import android.annotation.TargetApi;
 import android.content.Context;
-import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.Build;
-import android.support.annotation.ColorRes;
-import android.support.annotation.DrawableRes;
-import android.support.annotation.NonNull;
 import android.text.TextUtils;
 import android.view.View;
 
@@ -219,22 +215,6 @@ public class ReaderUtils {
      */
     public static ReaderTag getDefaultTag() {
         return getTagFromTagName(ReaderTag.TAG_TITLE_DEFAULT, ReaderTagType.DEFAULT);
-    }
-
-    /*
-     * returns a color resource - avoids having deprecation warnings everywhere getColor() is used
-     */
-    @SuppressWarnings("deprecation")
-    public static int getColorResource(@NonNull Context context, @ColorRes int colorResId) {
-        return context.getResources().getColor(colorResId);
-    }
-
-    /*
-     * returns a color resource - avoids having deprecation warnings everywhere getDrawable() is used
-     */
-    @SuppressWarnings("deprecation")
-    public static Drawable getDrawableResource(@NonNull Context context, @DrawableRes int drawableResId) {
-        return context.getResources().getDrawable(drawableResId);
     }
