+++ WordPress/src/main/java/org/wordpress/android/ui/reader/views/ReaderIconCountView.java
@@ -10,6 +10,7 @@ import android.widget.TextView;
 import org.wordpress.android.R;
 import org.wordpress.android.ui.reader.utils.ReaderUtils;
 import org.wordpress.android.util.FormatUtils;
+import org.wordpress.android.util.ResourceUtils;
 
 /*
  * used when showing comment + comment count, like + like count
@@ -53,10 +54,10 @@ public class ReaderIconCountView extends LinearLayout {
                 mIconType = a.getInteger(R.styleable.ReaderIconCountView_readerIcon, ICON_LIKE);
                 switch (mIconType) {
                     case ICON_LIKE :
-                        mImageView.setImageDrawable(ReaderUtils.getDrawableResource(context, R.drawable.reader_button_like));
+                        mImageView.setImageDrawable(ResourceUtils.getDrawableResource(context, R.drawable.reader_button_like));
                         break;
                     case ICON_COMMENT :
-                        mImageView.setImageDrawable(ReaderUtils.getDrawableResource(context, R.drawable.reader_button_comment));
+                        mImageView.setImageDrawable(ResourceUtils.getDrawableResource(context, R.drawable.reader_button_comment));
                         break;
                 }
